Merge maven profiles build-bundle and importable-projects
Since we use m2e (Maven Eclipse Integration), we do not need to
generate projects. This solution also fixes some of the problems
with the groovy script (See #5).

This commit was strongly inspired by the work of @hornm and @dscho on the
wip branch of this repository (which may have been removed by the time you
read this commit message).